isolate integration tests in separate package to avoid collisions in IDEA
