+++ generator-assertions-tests/pom.xml
@@ -37,7 +37,7 @@
                     <legacyClasses>
                         <param>io.stubbs.truth.generator.testing.legacy.NonBeanLegacy</param>
                     </legacyClasses>
-                    <entryPointClassPackage>io.stubbs.truth.generator</entryPointClassPackage>
+                    <entryPointClassPackage>io.stubbs.truth.generator.integrationTests</entryPointClassPackage>
                 </configuration>
             </plugin>
             <plugin>
