+++ generator-assertions-tests/src/test/java/io/stubbs/truth/generator/integrationTests/InstanceUtils.java
@@ -1,4 +1,4 @@
-package io.stubbs.truth.generator;
+package io.stubbs.truth.generator.integrationTests;
 
 public class InstanceUtils {
     static <T> T createInstance(Class<T> clazz) {
