+++ generator-assertions-tests/src/test/java/io/stubbs/truth/generator/integrationTests/capturedtemplates/MyIdCardSubject.java
@@ -1,9 +1,10 @@
-package io.stubbs.truth.generator.capturedtemplates;
+package io.stubbs.truth.generator.integrationTests.capturedtemplates;
 
 import com.google.common.truth.FailureMetadata;
 import io.stubbs.truth.generator.SubjectFactoryMethod;
 import io.stubbs.truth.generator.UserManagedMiddleSubject;
 import io.stubbs.truth.generator.UserManagedTruth;
+import io.stubbs.truth.generator.integrationTests.UserManagedSubjectsTest;
 import io.stubbs.truth.generator.internal.model.MiddleClass;
 import io.stubbs.truth.generator.testModel.IdCard;
 import io.stubbs.truth.generator.testModel.IdCardParentSubject;
@@ -13,7 +14,7 @@ import javax.annotation.processing.Generated;
 /**
  * Test user managed {@link MiddleClass} to ensure it is picked up and used in the graph.
  *
- * @see io.stubbs.truth.generator.UserManagedSubjectsTest
+ * @see UserManagedSubjectsTest
  */
 @UserManagedTruth(IdCard.class)
 @Generated(value = "io.stubbs.truth.generator.internal.TruthGenerator", date = "1970-01-01T00:00:00Z")
